Check for a merge if possible

+++ src/main/java/com/iamneo/springboot/LMS/model/Answer.java
@@ -1,21 +1,13 @@
 package com.iamneo.springboot.LMS.model;
 
-
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonBackReference;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 @Entity
 @Data
@@ -29,7 +21,7 @@ public class Answer {
     private long answerId;
 
     private String studentId;
-    private int marks;
+    private float marks;
 
     @JsonBackReference
     @OneToOne
